Terminei todas as funcionalidades obrigatórias, faltando ajustar layout e refatorar o código.
