+++ semana10/aula1-react-router/src/reducers/trips.js
@@ -1,5 +1,7 @@
 const initialState = {
     allTrips: [],
+    selectedTripId: "",
+    tripDetails: {},
 }
 
 const trips = (state = initialState, action) => {
@@ -7,6 +9,12 @@ const trips = (state = initialState, action) => {
         case 'SET_ALL_TRIPS':
             const tripList = action.payload.allTrips;
             return {...state, allTrips: tripList};
+        case 'SET_SELECTED_TRIP':
+            const tripId = action.payload.selectedTripId;
+            return {...state, selectedTripId: tripId}    
+        case 'SET_TRIP_DETAILS':
+            const tripDetails = action.payload.tripDetails;
+            return {...state, tripDetails: tripDetails}
         default:
             return state;
     }
